Finish separating container with tab
