+++ src/app/review/StudyProgress.js
@@ -2,7 +2,6 @@ import React from "react";
 import cx from "classnames";
 
 const StudyProgress = ({ index, items, pageSize, pageStart, pageEnd, isFinished, correctness }) => {
-  console.log("StudyProgress", index, items);
   return (
     <div className="d-flex align-items-center">
       {items.slice(pageStart, pageEnd).map((el, key) => (
